controller advancement

+++ controller/CartController.js
@@ -1,8 +1,7 @@
+const ensureAuthorization = require("../auth");
 const jwt = require("jsonwebtoken");
 const conn = require("../mariadb"); //db모듈
 const { StatusCodes } = require("http-status-codes");
-const dotenv = require("dotenv");
-dotenv.config();
 
 const addToCart = (req, res) => {
   const { book_id, quantity } = req.body;
@@ -42,8 +41,13 @@ const getCartItems = (req, res) => {
     });
   } else {
     let sql = `SELECT cartItems.id,book_id,title,summary,quantity,price from cartItems
-LEFT JOIN books ON cartItems.book_id = books.id WHERE user_id=? AND cartItems.id IN(?) `;
-    let values = [authorization.id, selected];
+LEFT JOIN books ON cartItems.book_id = books.id WHERE user_id=?  `;
+    let values = [authorization.id];
+    if (selected) {
+      //주문서 작성 시 '선택한 장바구니 목록 조회'
+      sql += `AND cartItems.id IN(?)`;
+      values.push(selected);
+    }
     conn.query(sql, values, (err, results) => {
       if (err) {
         console.log(err);
@@ -55,31 +59,27 @@ LEFT JOIN books ON cartItems.book_id = books.id WHERE user_id=? AND cartItems.id
 };
 
 const removeToCartItem = (req, res) => {
-  const { id } = req.params;
-  let sql = "DELETE FROM cartItems WHERE id = ?";
-  conn.query(sql, id, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
-};
-
-function ensureAuthorization(req, res) {
-  try {
-    let receivedJwt = req.headers["authorization"];
-    console.log("received jswt: ", receivedJwt);
-
-    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-    console.log(decodedJwt);
-    return decodedJwt;
-  } catch (err) {
-    console.log(err.name);
-    console.log(err.message);
-    return err;
+  let authorization = ensureAuthorization(req, res);
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토큰입니다 다시 로그인 하세요",
+    });
+  } else {
+    const cartItemId = req.params.id;
+    let sql = "DELETE FROM cartItems WHERE id = ?";
+    conn.query(sql, cartItemId, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
   }
-}
+};
 
 module.exports = {
   addToCart,
